refactor(dashboard): drop dead mission-utils re-export hub, point callers at lib direct
After iter#36 removed the `relativeTime` wrapper, `mission-utils.ts`
collapsed to a 6-line 100% pass-through file:

    import { formatDuration } from '../lib/format-time'
    import { trimText } from '../lib/truncate'
    import { toneClass } from '../lib/tone'
    import { statusLabel } from '../lib/status-label'
    export { formatDuration, trimText, toneClass, statusLabel }

No own logic, no test file, no remaining wrappers — just a re-export
layer that adds an extra module-graph hop and produces dual-path
imports: some callers go through this hub, others import from `lib/*`
directly, drifting on what counts as the "canonical" path. The hub's
`toneClass` and `statusLabel` re-exports have zero consumers (every
caller of those two already goes via `lib/`), so they were already
dead.

Delete the hub, repoint the five remaining consumers
(`keeper-detail-alert-strip`, `keeper-cognition-inspector`,
`agent-profile`, `agent-monitor/runtime-strip`, `agent-roster`) at
`lib/format-time` / `lib/truncate` directly. Update the
`vi.mock('../mission-utils')` in `runtime-strip.test.ts` to mock
`../../lib/format-time` instead — without this the mock silently
no-ops and the production `formatDuration` runs in the test
("renders activity when ageSeconds present" failed during verification
until the mock path was fixed; baseline 10/10 → 1 fail → 10/10 after).

iter#36 deleted a wrapper, iter#38 collects the hub itself — chained
dead-code detection.
