+++ dashboard/src/components/agent-roster.ts
@@ -30,7 +30,8 @@ import { AgentPresence } from './common/agent-presence'
 import { AgentCapability } from './common/agent-capability'
 import { openAgentDetail } from './agent-detail-state'
 import { openKeeperDetail } from './keeper-detail'
-import { formatDuration, trimText } from './mission-utils'
+import { formatDuration } from '../lib/format-time'
+import { trimText } from '../lib/truncate'
 import { formatTokens } from '../lib/format-number'
 import { namespaceTruth } from '../namespace-truth-store'
 import {
